yoke_derive

Module lifetimes

Functions

custom_lt

ignored_lifetime_ident

The returned ident should not be used in a way that impacts the code generated by the derive.

replace_lifetime

lt_param is the un-raw version of the lifetime parameter of the yokeable type. This function replaces all (possibly raw) occurrences of lt_param in the type x with the lt lifetime.

static_lt